----
I want a fastapi program which uses YouTube-DL (or the fork yt-dlp) as a python module (Embedded mode).
It should provide the following 3 routes:

1. List of providers:
A endpoint returning Domain names as a list for all sites supported by yt-dlp.

2. Download/Stream
This endpoint should initialise the download of an url and then stream a mp4 or mp3 or whatever format will be selected by input data.
If it is not yet available it should keep the connection open, and redirect to self if needed, so that a webbrowser would eventually just end up with the file.
Note, the file should have thumbnails, subtitles and meta information embedded. As much information as possible.

2. Meta about url:
Given a url, e.g. a YouTube URL, it should provide all the know meta information about it, like title, description, tags, times, and video data (length, etc.).
Additionally it should provide urls to the second endpoints.
For videos it should provide proper links in mkv (embedding everything we can embed), webm, mp4, mp3 (audio), and a png/jpg thumbnail.